review: screen a cluster proposal's structure before it is unioned
The blocking review finding: `verifiableClusters` screened only unknown-id
/ no-anchor / already-clustered, so a structurally invalid proposal still
unioned its members into the group. A cross-path member pulled in that way
can out-rank the real survivor, and the tier-1 merge underneath then
collapses to nothing -- recorded in neither `merges` nor
`clusterRejections`. Tier 2 could degrade BELOW tier 1, which is the one
thing dispatch-cluster.ts promises it cannot.

The path/source/severity rules now run per proposal before any union, on
the anchor the proposal's own members elect (its blocking member if it has
one, else the first the id screen kept -- only the severity rule is
asymmetric, and 'advisory absorbed into blocking' must stay legal). What
survives is a membership hint whose members are mutually absorbable; the
per-member re-check against the ACTUAL survivor stays, because tier-1
bridging can still elect a claim the proposal never named.

Two tests: the lost-merge shape itself (fails without the screen), and the
survivor-time re-check that the screen does not subsume -- a member legal
against the anchor but same-source as the survivor tier 1 elects.

Also splits the modules the merge with main pushed over the 1000-line cap:
tier 2 to lib/dedup-cluster.ts (the name its tests already carried) and the
eval's dedup stage to eval/live-dedup.ts, mirroring dispatch-cluster.ts on
each side. dispatch.ts sheds the empty section header main's dispatch-agents
split left behind, and dedup.ts's orphaned `mergeable` doc comment is
reattached to `mergeable`.

Plus the review's other import ask: live-producer.ts takes the exported
CLUSTERER constant rather than restating the agent name.
